Mark more integer API schema fields as integers

+++ packages/lms-external-backend-interfaces/src/llmBackendInterface.ts
@@ -73,7 +73,7 @@ export function createLlmBackendInterface() {
         inputString: z.string(),
       }),
       returns: z.object({
-        tokenCount: z.number(),
+        tokenCount: z.number().int(),
       }),
     });
 }
